A bunch of new features

+++ src/api/groups.ts
@@ -1,17 +1,18 @@
 import { GroupType } from "../types/GroupType";
 import { api } from ".";
-import { getData } from "../storage";
 
-
-export async function getAllGroups(): Promise<GroupType[]> {
-    return api.get(`/groups/`)
-        .then(res => {
-            return res.data.groups
-        })
-        .catch(err => {
-            console.error(JSON.stringify(err));
-            return []
-        })
+export async function getAllGroups(token: string): Promise<GroupType[]> {
+    return api.get(`/groups/`, {
+        headers: {
+            Authorization: token
+        }
+    }).then(res => {
+        return res.data.groups
+    })
+    .catch(err => {
+        console.error(JSON.stringify(err));
+        return []
+    })
 }
 
 type ICreateGroup = {
@@ -19,16 +20,35 @@ type ICreateGroup = {
     description: string,
     usersEmails: string[]
 }
-export async function createGroup({name, description, usersEmails}: ICreateGroup): Promise<GroupType> {
-    console.log(usersEmails)
+export async function createGroup(token: string, {name, description, usersEmails}: ICreateGroup): Promise<GroupType> {
     return api.post(`/groups`, {
-        userId: getData("user").id,
         group: {
             name,
             description,
         },
-        users: usersEmails
+        users: usersEmails || []
+    }, {headers: {Authorization: token}})
+    .then(res => res.data)
+    .catch(err => {
+        console.error(err);
+        return null
     })
+}
+
+type IUpdateGroup = {
+    id: string,
+    name: string,
+    description: string,
+    usersEmails: string[]
+}
+export async function updateGroup(token: string, {id, name, description, usersEmails}: IUpdateGroup): Promise<GroupType> {
+    return api.patch(`/groups/${id}`, {
+        group: {
+            name,
+            description,
+        },
+        users: usersEmails || []
+    }, {headers: {Authorization: token}})
     .then(res => res.data)
     .catch(err => {
         console.error(err);
@@ -37,7 +57,7 @@ export async function createGroup({name, description, usersEmails}: ICreateGroup
 
 }
 
-export async function deleteGroup(id: string): Promise<void> {
-    return api.delete(`/groups/${id}`)
+export async function deleteGroup(token: string, id: string): Promise<void> {
+    return api.delete(`/groups/${id}`, {headers: {Authorization: token}})
 }
